requirements.txt
torch
transformers
sentence-transformers
nltk
PyPDF2
python-docx
tqdm
